third_party/blink/web_tests/http/tests/devtools/debugger/fetch-breakpoints-expected.txt

Tests fetch() breakpoints.


Running: testFetchBreakpoint
Script execution paused.
Call stack:
    0) sendRequest (fetch-breakpoints.js:18)
    1)  (:1)
Script execution resumed.

Running: testPauseOnAnyFetch
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.

Running: testDisableBreakpoint
Script execution paused.
Script execution resumed.
